+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new token **{mi_audio_sample_rate}**, this token will provide audio track #1's sampling rate (48 kHz).
 - Added new token **{mi_audio_bitrate}**, this will provide audio track #1's bitrate (640000).
 - Added new token **{mi_audio_bitrate_formatted}**, this will provide audio track #1's bitrate (640 kb/s).
+- Added new token **{mi_video_width}**, this will provide video track #1's width (1940).
+- Added new token **{mi_video_height}**, this will provide video track #1's height (1080).
 
 ### Changed